+++ rl/examples/factory.py
@@ -8,10 +8,10 @@ from rl import print_exc
 
 
 @print_exc
+@generator
 def complete_command(text):
     # Return executables matching 'text'
     for dir in os.environ.get('PATH').split(':'):
-        dir = os.path.expanduser(dir)
         if os.path.isdir(dir):
             for name in os.listdir(dir):
                 if name.startswith(text):
@@ -21,7 +21,7 @@ def complete_command(text):
 
 def main():
     # Set the completion entry function
-    completer.completer = generator(complete_command)
+    completer.completer = complete_command
 
     # Enable TAB completion
     completer.parse_and_bind('TAB: complete')