+++ rl/examples/static.py
@@ -9,6 +9,7 @@ strings = ['foo-', 'bar-', 'baz+', 'peng+']
 
 
 @print_exc
+@generator
 def complete_string(text):
     completion.suppress_append = True
     return [x for x in strings if x.startswith(text)]
@@ -16,7 +17,7 @@ def complete_string(text):
 
 def main():
     # Set the completion entry function
-    completer.completer = generator(complete_string)
+    completer.completer = complete_string
 
     # Enable TAB completion
     completer.parse_and_bind('TAB: complete')