+++ rl/tests/test_callhooks.py
@@ -187,11 +187,10 @@ class DisplayMatchesHookTests(JailSetup):
         self.assertEqual(called, [('fred.', ['fred.gif', 'fred.txt'], 8)])
 
     # FIXME: Prints to rl_outstream
-    def DISABLED_test_default_display_matches_hook(self):
+    def test_default_display_matches_hook(self):
         self.mkfile('fred.txt', 'fred.gif')
         completer.completer = filecomplete
         readline.complete_internal('?')
-        self.assertEqual(called, [])
 
 
 class WordBreakHookTests(unittest.TestCase):
@@ -373,6 +372,35 @@ class DirectoryCompletionHookTests(JailSetup):
         readline.complete_internal(TAB)
         self.assertEqual(completion.line_buffer, "flint\\ stone/fr")
 
+    # FIXME: Prints to rl_outstream
+    def test_called_by_display_match_list(self):
+        def func(dirname):
+            called.append(dirname)
+            return dirname.replace('\\', '')
+        self.mkdir('flint stone')
+        self.mkfile('flint stone/fred.txt')
+        if readline.readline_version() >= 0x0801:
+            # Reset last_completion_failed
+            readline.complete_internal(TAB)
+        completer.directory_completion_hook = func
+        completion.line_buffer = 'flint\\ stone/fr'
+        readline.complete_internal('?')
+        self.assertEqual(called, ['flint\\ stone/', 'flint stone'])
+        self.assertEqual(completion.line_buffer, "flint\\ stone/fr")
+
+    def test_preempted_by_directory_rewrite_hook(self):
+        def func(dirname):
+            called.append(dirname)
+            return dirname.replace('\\', '')
+        self.mkdir('flint stone')
+        self.mkfile('flint stone/fred.txt')
+        completer.directory_rewrite_hook = lambda x: x
+        completer.directory_completion_hook = func
+        completion.line_buffer = 'flint\\ stone/fr'
+        readline.complete_internal(TAB)
+        self.assertEqual(called, []) # Not called
+        self.assertEqual(completion.line_buffer, "flint\\ stone/fr")
+
 
 class FilenameDequotingFunctionTests(JailSetup):
 
@@ -418,7 +446,10 @@ class FilenameDequotingFunctionTests(JailSetup):
         completer.filename_dequoting_function = func
         completion.line_buffer = 'fr\\ '
         readline.complete_internal(TAB)
-        self.assertEqual(called, [('.', ''), ('fr\\ ', '')])
+        if readline.readline_version() >= 0x0801:
+            self.assertEqual(called, [('.', ''), ('fr\\ ', ''), ('fr\\ ', '')])
+        else:
+            self.assertEqual(called, [('.', ''), ('fr\\ ', '')])
         self.assertEqual(completion.line_buffer, "'fr ed.txt' ")
 
     def test_no_dequoting(self):
@@ -440,7 +471,10 @@ class FilenameDequotingFunctionTests(JailSetup):
         completer.filename_dequoting_function = func
         completion.line_buffer = "'fr "
         readline.complete_internal(TAB)
-        self.assertEqual(called, [('.', "'"), ('fr ', "'")])
+        if readline.readline_version() >= 0x0801:
+            self.assertEqual(called, [('.', "'"), ('fr ', "'"), ('fr ', "'")])
+        else:
+            self.assertEqual(called, [('.', "'"), ('fr ', "'")])
         self.assertEqual(completion.line_buffer, "'fr ed.txt' ")
 
     def test_empty_string(self):
@@ -800,4 +834,31 @@ class FilenameStatHookTests(JailSetup):
         readline.complete_internal(TAB)
         self.assertEqual(called, ['flintstone/fred.txt'])
         self.assertEqual(completion.line_buffer, "flintstone/fred.txt ")
+
+    # FIXME: Prints to rl_outstream
+    def test_called_by_display_match_list(self):
+        def func(filename):
+            called.append(filename)
+            return filename
+        self.mkdir('flintstone')
+        self.mkfile('flintstone/fred.txt')
+        completer.filename_stat_hook = func
+        completion.line_buffer = 'flintstone/fr'
+        readline.complete_internal('?')
+        self.assertEqual(called, ['flintstone/fred.txt'])
+        self.assertEqual(completion.line_buffer, "flintstone/fr")
+
+    # FIXME: Prints to rl_outstream
+    def test_preempted_by_directory_completion_hook(self):
+        def func(filename):
+            called.append(filename)
+            return filename
+        self.mkdir('flintstone')
+        self.mkfile('flintstone/fred.txt')
+        completer.directory_completion_hook = lambda x: x
+        completer.filename_stat_hook = func
+        completion.line_buffer = 'flintstone/fr'
+        readline.complete_internal('?')
+        self.assertEqual(called, []) # Not called
+        self.assertEqual(completion.line_buffer, "flintstone/fr")